requirements.txt
aiogram==3.4.1
pymongo==4.6.2
motor==3.3.2
python-dateutil==2.8.2